* master:
  bump sdk version to v0.43.0-rc2
  Implement reward distribution (#25)
  Add issue, pr templates

# Conflicts:
#	x/farming/keeper/msg_server.go
#	x/farming/keeper/plan.go
#	x/farming/keeper/plan_test.go
#	x/farming/keeper/proposal_handler.go
#	x/farming/types/errors.go
#	x/farming/types/plan.go

+++ x/farming/keeper/staking.go
@@ -3,8 +3,10 @@ package keeper
 import (
 	"encoding/binary"
 
-	sdk "github.com/cosmos/cosmos-sdk/types"
+	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	gogotypes "github.com/gogo/protobuf/types"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	"github.com/tendermint/farming/x/farming/types"
 )
@@ -111,7 +113,9 @@ func (k Keeper) DeleteStaking(ctx sdk.Context, staking types.Staking) {
 	store := ctx.KVStore(k.storeKey)
 	store.Delete(types.GetStakingKey(staking.Id))
 	store.Delete(types.GetStakingByFarmerIndexKey(staking.GetFarmer()))
-	k.DeleteStakingCoinDenomsIndex(ctx, staking.Id, staking.StakingCoinDenoms())
+	if denoms := staking.StakingCoinDenoms(); len(denoms) > 0 {
+		k.DeleteStakingCoinDenomsIndex(ctx, staking.Id, denoms)
+	}
 }
 
 // DeleteStakingCoinDenomsIndex removes an staking for the staking mapper store.
@@ -175,22 +179,21 @@ func (k Keeper) ReleaseStakingCoins(ctx sdk.Context, farmer sdk.AccAddress, unst
 }
 
 // Stake stores staking coins to queued coins and it will be processed in the next epoch.
-func (k Keeper) Stake(ctx sdk.Context, farmer sdk.AccAddress, amount sdk.Coins) error {
+func (k Keeper) Stake(ctx sdk.Context, farmer sdk.AccAddress, amount sdk.Coins) (types.Staking, error) {
 	if err := k.ReserveStakingCoins(ctx, farmer, amount); err != nil {
-		return err
+		return types.Staking{}, err
 	}
 
 	params := k.GetParams(ctx)
 
 	staking, found := k.GetStakingByFarmer(ctx, farmer)
 	if !found {
-		farmingFeeCollectorAcc, err := sdk.AccAddressFromBech32(params.FarmingFeeCollector)
-		if err != nil {
-			return err
-		}
+		// We don't need to check error here because it is already checked
+		// in Params validation.
+		farmingFeeCollectorAcc, _ := sdk.AccAddressFromBech32(params.FarmingFeeCollector)
 
 		if err := k.bankKeeper.SendCoins(ctx, farmer, farmingFeeCollectorAcc, params.StakingCreationFee); err != nil {
-			return err
+			return types.Staking{}, sdkerrors.Wrap(types.ErrFeeCollectionFailure, err.Error())
 		}
 
 		staking = k.NewStaking(ctx, farmer)
@@ -200,26 +203,18 @@ func (k Keeper) Stake(ctx sdk.Context, farmer sdk.AccAddress, amount sdk.Coins) 
 	k.SetStaking(ctx, staking)
 	k.SetStakingIndex(ctx, staking)
 
-	ctx.EventManager().EmitEvents(sdk.Events{
-		sdk.NewEvent(
-			types.EventTypeStake,
-			sdk.NewAttribute(types.AttributeKeyFarmer, farmer.String()),
-			sdk.NewAttribute(types.AttributeKeyStakingCoins, amount.String()),
-		),
-	})
-
-	return nil
+	return staking, nil
 }
 
 // Unstake unstakes an amount of staking coins from the staking reserve account.
-func (k Keeper) Unstake(ctx sdk.Context, farmer sdk.AccAddress, amount sdk.Coins) error {
+func (k Keeper) Unstake(ctx sdk.Context, farmer sdk.AccAddress, amount sdk.Coins) (types.Staking, error) {
 	staking, found := k.GetStakingByFarmer(ctx, farmer)
 	if !found {
-		return types.ErrStakingNotExists
+		return types.Staking{}, types.ErrStakingNotExists
 	}
 
 	if err := k.ReleaseStakingCoins(ctx, farmer, amount); err != nil {
-		return err
+		return types.Staking{}, err
 	}
 
 	prevDenoms := staking.StakingCoinDenoms()
@@ -239,6 +234,7 @@ func (k Keeper) Unstake(ctx sdk.Context, farmer sdk.AccAddress, amount sdk.Coins
 
 	// Remove the Staking object from the kvstore when all coins has been unstaked
 	// and there's no rewards left.
+	// TODO: find more efficient way to check if the farmer has no rewards
 	if staking.StakedCoins.IsZero() && staking.QueuedCoins.IsZero() && len(k.GetRewardsByFarmer(ctx, farmer)) == 0 {
 		k.DeleteStaking(ctx, staking)
 	} else {
@@ -256,16 +252,23 @@ func (k Keeper) Unstake(ctx sdk.Context, farmer sdk.AccAddress, amount sdk.Coins
 			}
 		}
 
-		k.DeleteStakingCoinDenomsIndex(ctx, staking.Id, removedDenoms)
+		if len(removedDenoms) > 0 {
+			k.DeleteStakingCoinDenomsIndex(ctx, staking.Id, removedDenoms)
+		}
 	}
 
-	ctx.EventManager().EmitEvents(sdk.Events{
-		sdk.NewEvent(
-			types.EventTypeUnstake,
-			sdk.NewAttribute(types.AttributeKeyFarmer, farmer.String()),
-			sdk.NewAttribute(types.AttributeKeyUnstakingCoins, amount.String()),
-		),
-	})
+	// We're returning a Staking even if it has deleted.
+	return staking, nil
+}
 
-	return nil
+// ProcessQueuedCoins moves queued coins into staked coins.
+func (k Keeper) ProcessQueuedCoins(ctx sdk.Context) {
+	k.IterateAllStakings(ctx, func(staking types.Staking) (stop bool) {
+		if !staking.QueuedCoins.IsZero() {
+			staking.StakedCoins = staking.StakedCoins.Add(staking.QueuedCoins...)
+			staking.QueuedCoins = sdk.NewCoins()
+			k.SetStaking(ctx, staking)
+		}
+		return false
+	})
 }
